[FEAT]: 호가 단위에 맞게 금액을 제시하는지 확인하는 기능 구현
