+++ pickme/src/main/resources/static/css/bid/bidPage.css
@@ -98,7 +98,7 @@ h2 {
 
 @keyframes growShrink {
     0% { transform: scale(1); }
-    50% { transform: scale(1.5); }
+    50% { transform: scale(2.5); }
     100% { transform: scale(1); }
 }
