+++ pickme/src/main/resources/static/css/bid/modal.css
@@ -23,9 +23,9 @@
 
 .modal-header {
     display: flex;
-    justify-content: center; /* 수평 중앙 정렬 */
-    align-items: center; /* 수직 중앙 정렬 */
-    height: 100px; /* 필요에 따라 조정 */
+    justify-content: center;
+    align-items: center;
+    height: 100px;
 }
 
 .icon-container {
