+++ pickme/src/main/resources/templates/bid/bidPage.html
@@ -72,7 +72,6 @@
 </script>
 <script type="module" src="/js/bid/modal.js"></script>
 <script src="/js/bid/timer.js"></script>
-<script src="/js/bid/money.js"></script>
-<script type="module" src="/js/bid/socket.js"></script>
+<script type="module" src="/js/bid/bid.js"></script>
 </body>
 </html>
